% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmCalc.R
\name{dmCalcServer}
\alias{dmCalcServer}
\title{DM单计算}
\usage{
dmCalcServer(input, output, session, dms_token)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{dms_token}{口令}
}
\value{
返回值
}
\description{
DM单计算
}
\examples{
dmCalcServer()
}
