% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lc.R
\name{downloadserver}
\alias{downloadserver}
\title{Title 下载日志}
\usage{
downloadserver(input, output, session, dms_token)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{dms_token}{口令}
}
\value{
返回值
}
\description{
Title 下载日志
}
\examples{
downloadserver()
}
