% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priceDetail.R
\name{priceDetailServer}
\alias{priceDetailServer}
\title{DM单计算}
\usage{
priceDetailServer(input, output, session, dms_token)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{dms_token}{口令}
}
\value{
返回值
}
\description{
DM单计算
}
\examples{
priceDetailServer()
}
