% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priceModel.R
\name{priceModelServer}
\alias{priceModelServer}
\title{查询价格查询注册函数}
\usage{
priceModelServer(input, output, session, dms_token)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{dms_token}{口令}
}
\value{

}
\description{
查询价格查询注册函数
}
\examples{
priceModelServer()
}
